+++ NEWS.md
@@ -2,6 +2,7 @@ Development version
 
 * switch group and unit definitions, to be consistent with the literature
 * add Theil's Information Index (H) 
+* add entropy function
 
 v0.0.1
 ======